movie_app_backend/requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djongo==1.2.31
dnspython==2.7.0
pillow==11.2.1
pymongo==4.12.0
python-dotenv==1.1.0
sqlparse==0.5.3
